Health Script fixed

+++ GAMPRG3_2DHellOnEarth/Assets/Scripts/Player/HealthBar.cs
@@ -7,13 +7,13 @@ public class HealthBar : MonoBehaviour
 {
     public Slider HealthSlider;
 
-    public void SetMaxHealth(int Health)
+    public void SetMaxHealth(float Health)
     {
         HealthSlider.maxValue = Health;
         HealthSlider.value = Health;
     }
 
-    public void SetHealth(int Health)
+    public void SetHealth(float Health)
     {
         HealthSlider.value = Health;
     }
